fix where flattening starts nesting wise

+++ go/flatten.go
@@ -7,11 +7,12 @@ import "log"
 import "os"
 import "path"
 import "regexp"
+import "strings"
 import "syscall"
 import "github.com/otiai10/copy"
 
 var scopedModuleRe,_ = regexp.Compile(`node_modules/@([^/])*/([^/])*$`)
-var stdModuleRe = regexp.MustCompile(`node_modules/[^\.^@^/.]+$`)
+var stdModuleRe,_ = regexp.Compile(`node_modules/[^\.^@^/.]*$`)
 
 func IsSyml(fi fs.FileInfo) bool {
 	if fi != nil && fi.Mode() & os.ModeSymlink != 0 {
@@ -46,24 +47,53 @@ func CanWrite(path string) bool {
 	}
 }
 
+func stringInSlice(a string, list []string) bool {
+    for _, b := range list {
+        if b == a {
+            return true
+        }
+    }
+    return false
+}
+
+func countMatches(s string, re *regexp.Regexp) int {
+    total := 0
+    for start := 0; start < len(s); {
+        remaining := s[start:] // slicing the string is cheap
+        loc := re.FindStringIndex(remaining)
+        if loc == nil {
+            break
+        }
+        // loc[0] is the start index of the match,
+        // loc[1] is the end index (exclusive)
+        start += loc[1]
+        total++
+    }
+    return total
+}
+
+
 func NodeModuleDirs(root string, sys fs.FS) ([]string, []string, error) {
 	var scoped []string
 	var standard []string
+	nodeModP := regexp.MustCompile("node_modules")
+
 
 	err := fs.WalkDir(sys, root, func(path string, de fs.DirEntry, err error) error {
 		if err != nil {
 			return err
 		}
+		// nodeModPM := nodeModP.FindAllStringIndex(path, -1)
+		if countMatches(path, nodeModP) > 1 && !strings.HasSuffix(path, "node_modules") {
+			var scopedResult = scopedModuleRe.MatchString(path)
+			if  scopedResult {
+				scoped = append(scoped, path)
+			}
 
-		var scopedResult = scopedModuleRe.MatchString(path)
-		if  scopedResult {
-			scoped = append(scoped, path)
+			if stdModuleRe.MatchString(path) {
+				standard = append(standard, path)
+			}
 		}
-
-		if stdModuleRe.MatchString(path) {
-			standard = append(standard, path)
-		}
-
 		return nil
 	})
 
@@ -73,9 +103,18 @@ func NodeModuleDirs(root string, sys fs.FS) ([]string, []string, error) {
 
 func main() {
 	if PathExists("node_modules") {
+		var topLevelDeps []string
+		topLevelDepsInfo, lsErr := ioutil.ReadDir("node_modules")
+		if lsErr != nil {
+			log.Fatal(lsErr)
+		}
+		for _, topFile := range topLevelDepsInfo {
+			topLevelDeps = append(topLevelDeps, topFile.Name())
+		}
 		pwd, _ := os.Getwd()
 		nfs := os.DirFS(pwd)
 		scoped, standard, _ := NodeModuleDirs("node_modules", nfs)
+
 
 		for _, stdFile := range standard {
 			syminfo, _ := os.Lstat(stdFile)
@@ -103,7 +142,9 @@ func main() {
 			dirName := path.Base(stdFile)
 			if _, err := os.Stat(path.Join("./node_modules", dirName)); err == nil {
 				// path exists
-				os.RemoveAll(stdFile)
+				if stringInSlice(dirName, topLevelDeps) {
+					os.RemoveAll(stdFile)
+				}
 
 			} else if os.IsNotExist(err) {
 				// path does *not* exist
@@ -139,10 +180,12 @@ func main() {
 			}
 			dirName := path.Base(scopedFile)
 			scopeName := path.Base(path.Dir(scopedFile))
-			pName := path.Join(dirName, scopeName)
+			pName := path.Join(scopeName, dirName)
 			if _, err := os.Stat(path.Join("./node_modules", pName)); err == nil {
 				// path exists
-				os.RemoveAll(scopedFile)
+				if stringInSlice(pName, topLevelDeps) {
+					os.RemoveAll(scopedFile)
+				}
 
 			} else if os.IsNotExist(err) {
 				// path does *not* exist
